Encoder Check Update
Implemented the encoder check logic and the looping mechanism for
constant checks. Still need to have a timeout in between loops to
prevent it from accidentally looping to fast and killing it when it is
running slowly, and a canceling command to stop the launcher rising
actions when it is going wrong.