TRAJECTORY AND DISTANCE GET CHANGES
I have modified the trajectory table to change it into a regression
function method that returns the angle based on the distance inputed. I
also changed how we get the distances from the ultrasound and the
pixyCam, they now have their respective methods. We still need to
implement some logic that uses the pixel distance between the two upper
posts of the tape to a linear function that returns the distance from
the tower. And then use that to compare with the pixyCam distance and
the ultraSound to see which we should prioritize.