Tuned launcher values and created Launcher method to drop the launcher
The drop launcher method sets the launcher to 0 when the flightstick
trigger is pressed, and immediately raises it after it is released. This
allows drivers to do multiple obstacles including the cheval de frise
and the portcullis. The launcher setpoints were tuned, and the spin up
delay was lowered from 4 seconds to 2 seconds. I created a constant for
easy tuning as mechanical team predicts that the newer launcher motors
will spin up way faster. (Possibly one second or less). In addition,
some small issues were fixed, and the ultrasonic code was implemented.

NOTE: Inside of the Launcher class, the driver station report warning
was commented out, as my libraries failed to update correctly. Please
Uncomment. Thank you, Avidh